+++ CHANGELOG.md
@@ -13,3 +13,4 @@ Changelog can generally be found here https://www.spigotmc.org/resources/17067/u
 - Color codes from Discord will no longer make the message colored
 - Using #saveDefaultConfig, recommended to reset the configuration file due to major changes to the format
 - Sets the Bots game to "Minecraft"
+- Support nicknames for message templates and mentions